Move fetching code to Quotes container
