+++ src/containers/Quotes.js
@@ -2,6 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
+import Db from '../firebase';
+
 import ListItem from '../components/ListItem';
 import ListHeader from '../components/ListHeader';
 
@@ -12,8 +14,58 @@ const ListItemsContainer = styled.div`
 `;
 
 class Quotes extends Component {
+  constructor() {
+    super();
+
+    this.state = { quotes: [] };
+    this.deleteQuote = this.deleteQuote.bind(this);
+  }
+
+  componentDidMount() {
+    this.retrieveData();
+  }
+
+  componentWillUnmount() {
+    this.unsubscribe();
+  }
+
+  retrieveData() {
+    // TODO: Dynamic chat id
+    const chatId = '-394500082';
+    this.collectionRef = Db.collection('chats').doc(chatId).collection('msgs');
+    this.unsubscribe = this.collectionRef.orderBy('datetime').onSnapshot((res) => {
+      // res is QuerySnapshot, res.docs is an array of queryDocumentSnapshots
+      const quotesArr = res.docs.map((msgDoc) => {
+        const msgDocData = msgDoc.data();
+        let datetime = new Date();
+
+        // Check if key is available for use as datetime takes a while
+        if (msgDocData.hasOwnProperty('datetime') && msgDocData.datetime.hasOwnProperty('seconds')) {
+          datetime = new Date(msgDocData.datetime.seconds * 1000);
+        }
+
+        const { id } = msgDoc;
+        const { msg, user } = msgDocData;
+
+        return { id, msg, user, datetime };
+      });
+
+      this.setState({ quotes: quotesArr });
+      console.log('Refreshed data');
+    }, (err) => {
+      console.log(`Unable to refresh data: ${err}`);
+    });
+  }
+
+  deleteQuote(quoteId) {
+    this.collectionRef.doc(quoteId).delete()
+      .then(() => console.log('Quote successfully deleted'))
+      .catch(() => console.log('Unable to delete quote'));
+  }
+
   renderQuotes() {
-    const { startSN, quotes, deleteQuote } = this.props;
+    const { startSN } = this.props;
+    const { quotes } = this.state;
 
     return quotes.map((q, idx) => (
       <ListItem
@@ -22,7 +74,7 @@ class Quotes extends Component {
         sn={startSN + idx}
         quote={q.msg}
         username={q.user}
-        deleteQuote={deleteQuote}
+        deleteQuote={this.deleteQuote}
       />
     ));
   }
@@ -39,14 +91,8 @@ class Quotes extends Component {
   }
 }
 
-Quotes.defaultProps = {
-  quotes: []
-};
-
 Quotes.propTypes = {
-  deleteQuote: PropTypes.func.isRequired,
-  startSN: PropTypes.number.isRequired,
-  quotes: PropTypes.arrayOf(PropTypes.object)
+  startSN: PropTypes.number.isRequired
 };
 
 export default Quotes;
